refactored redux, combined relevant reducers
