+++ src/components/Item/sizeFilter.js
@@ -6,7 +6,7 @@ import {ReactComponent as DownArrow} from '../../assets/images/arrowDown.svg'
 export default function SizeFilter({ gender }) {
 
     const dispatch = useDispatch()
-    const size = useSelector(state => state.size)
+    const size = useSelector(state => state.item.size)
     const genderSymbol = (gender === 'men') ? 'M' : 'W'
 
     return (
