+++ src/components/Modals/sizeModal.js
@@ -12,7 +12,7 @@ export default function SizeModal({ gender }) {
     const wrapperRef = useRef(null)
     useOutsideAlerter(wrapperRef)
 
-    const size = useSelector(state => state.size)
+    const size = useSelector(state => state.item.size)
 
     const mensSneakerSizes = [7, 7.5, 8, 8.5, 9, 9.5, 10, 10.5, 11, 11.5, 12, 12.5, 13, 13.5, 14]
     const womensSneakerSizes = [5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 9.5, 10, 10.5, 11, 11.5, 12]
