--- src/redux/reducers/browseDataReducer.js
@@ -1,11 +1,0 @@
-
-const browseDataReducer = (state=[], action) => {
-    switch (action.type) {
-        case 'BROWSE_CALL':
-            return action.payload
-        default:
-            return state
-    }
-}
-
-export default browseDataReducer
