--- src/redux/reducers/itemPricesReducer.js
@@ -1,11 +1,0 @@
-
-const itemPricesReducer = (state=[], action) => {
-    switch(action.type) {
-        case 'UPDATE_ITEM_PRICES':
-            return action.payload
-        default:
-            return state
-    }
-}
-
-export default itemPricesReducer
