--- src/redux/reducers/searchModalReducer.js
@@ -1,12 +1,0 @@
-const searchModalReducer = (state = false, action) => {
-    switch(action.type) {
-        case 'SHOW_SEARCH_MODAL':
-            return true
-        case 'HIDE_SEARCH_MODAL':
-            return false
-        default:
-            return state
-    }
-}
-
-export default searchModalReducer
