--- src/redux/reducers/showFilterReducer.js
@@ -1,12 +1,0 @@
-const filterVisibleReducer = (state = false, action) => {
-    switch(action.type) {
-        case 'SHOW_FILTER':
-            return true
-        case 'HIDE_FILTER':
-            return false
-        default:
-            return state
-    }
-}
-
-export default filterVisibleReducer
