--- src/redux/reducers/sizeReducer.js
@@ -1,10 +1,0 @@
-const sizeReducer = (state=10, action) => {
-    switch(action.type) {
-        case 'UPDATE_SIZE':
-            return action.payload
-        default:
-            return state
-    }
-}
-
-export default sizeReducer
