--- src/redux/reducers/trendingReducer.js
@@ -1,10 +1,0 @@
-const trendingReducer = (state=[], action) => {
-    switch (action.type) {
-        case 'TRENDING_CALL':
-            return action.payload
-        default:
-            return state
-    }
-}
-
-export default trendingReducer
