--- src/redux/reducers/under200Reducer.js
@@ -1,10 +1,0 @@
-const under200Reducer = (state=[], action) => {
-    switch (action.type) {
-        case 'UNDER_200_CALL':
-            return action.payload
-        default:
-            return state
-    }
-}
-
-export default under200Reducer
